+++ master/CSharpToTypeScript.Test/TestData/src/CSharpToTypeScript.TestNamespace.TestMethod2.ts
@@ -1,0 +1,8 @@
+﻿import { Gender } from './CSharpToTypeScript.Test';
+import { RegistrationStatus } from './CSharpToTypeScript.TestNamespace.Enums';
+
+export const getGenderStatusClassName = (isValidated: boolean | undefined, status: RegistrationStatus, gender: Gender): string =>
+	gender !== Gender.Unknown && status !== RegistrationStatus.ApprovalPending ? status.toString() : (isValidated ? "approval-pending is-valid" : "approval-pending");
+
+export const getStatusClassName = (isValidated: boolean | undefined, status: RegistrationStatus): string =>
+	status !== RegistrationStatus.ApprovalPending ? status.toString() : (isValidated ? "approval-pending is-valid" : "approval-pending");